change migration fields integer by float

+++ priv/repo/migrations/20200228195852_create_football_player.exs
@@ -6,18 +6,18 @@ defmodule NflRushing.Repo.Migrations.CreateFootballPlayer do
       add :player_name, :string
       add :team_abbreviation, :string
       add :player_position, :string
-      add :rushing_attemps_per_game_avg, :integer
-      add :rushing_attemps, :integer
-      add :total_rushing_yards, :integer
-      add :rushing_average_yards_per_attempt, :integer
-      add :rushing_yards_per_game, :integer
-      add :total_rushing_touchdowns, :integer
-      add :longest_rush, :integer
-      add :rushing_first_downs, :integer
-      add :rushing_first_down_percentage, :integer
-      add :rushing_20_yards_each, :integer
-      add :rushing_40_yards_each, :integer
-      add :rushing_fumbles, :integer
+      add :rushing_attemps_per_game_avg, :float
+      add :rushing_attemps, :float
+      add :total_rushing_yards, :float
+      add :rushing_average_yards_per_attempt, :float
+      add :rushing_yards_per_game, :float
+      add :total_rushing_touchdowns, :float
+      add :longest_rush, :float
+      add :rushing_first_downs, :float
+      add :rushing_first_down_percentage, :float
+      add :rushing_20_yards_each, :float
+      add :rushing_40_yards_each, :float
+      add :rushing_fumbles, :float
 
       timestamps()
     end
